Algo1/requirements.txt
openai-whisper
googletrans==4.0.0-rc1
srt
tqdm
